renamed models according to angulars convention
