+++ src/app/model/admin-ingredient.model.spec.ts
@@ -1,4 +1,4 @@
-import { AdminIngredient } from './admin-ingredient';
+import { AdminIngredient } from './admin-ingredient.model';
 
 describe('AdminIngredient', () => {
   it('should create an instance', () => {
