+++ src/app/model/ingredient.model.spec.ts
@@ -1,4 +1,4 @@
-import { Ingredient } from './ingredient';
+import { Ingredient } from './ingredient.model';
 
 describe('Ingredient', () => {
   it('should create an instance', () => {
